feat: inflation on updateStatus/resolveChallenge
this commit enables inflation of a token's totalSupply while a
challenge is being resolved. the inflation is determined by a parameter
set by voters. the inflated tokens are available to be withdrawn by
voters who voted in the majority_bloc. the minimum deposit is then
increased to follow/support the inflation.

* remove functions to inflate/deflate supply in registry
* fix logic & move inflation code into scope (in resolveChallenge)
* add totalVotingTokens and tokenSupply states to Challenge structs
* rewrite tests: change supplyOracle in before block